Support non-integer exponents in Mandelbrot and Julia fractals with fractional-power paths

+++ wxChaos/core/fractals/JuliaZM.h
@@ -13,7 +13,7 @@
 class JuliaZM: public Fractal
 {
     JuliaZMRenderer* _myRender;
-    int _m;
+    double _m;
     double _bailout;
 public:
     JuliaZM(unsigned int width, unsigned int height);
